+++ caPutLogApp/caPutLogShellCommands.c
@@ -1,8 +1,17 @@
+#include <stdio.h>
+
 #include <dbDefs.h>
+#include <errlog.h>
 #include <iocsh.h>
 #include <epicsExport.h>
 
 #include "caPutLog.h"
+
+/* Use colored ERROR/WARNING text if available */
+#ifndef ERL_ERROR
+#  define ERL_ERROR "ERROR"
+#  define ERL_WARNING "WARNING"
+#endif
 
 static const iocshArg caPutLogInitArg0 = {"address", iocshArgString};
 static const iocshArg caPutLogInitArg1 = {"config", iocshArgInt};
@@ -46,15 +55,45 @@ static void caPutLogSetTimeFmtCall(const iocshArgBuf *args)
     caPutLogSetTimeFmt(args[0].sval);
 }
 
+/* Error message if caPutJsonLogInit used */
+static const iocshFuncDef caPutJsonLogInitDef = {"caPutJsonLogInit", 0, NULL};
+static void caPutJsonLogInitCall(const iocshArgBuf *args)
+{
+    fprintf(stderr, ERL_ERROR
+        ": The caPutLog module is configured for plain put-logging,\n"
+        "  only caPutLog* commands are available. To use the new JSON\n"
+        "  put-log format rebuild this IOC with 'caPutJsonLog.dbd'\n"
+        "  instead of 'caPutLog.dbd'.\n");
+    #ifdef IOCSHFUNCDEF_HAS_USAGE
+        iocshSetError(-1);
+    #endif
+}
+
 static void caPutLogRegister(void)
 {
-    static int done = FALSE;
+    extern int caPutLogRegisterDone;
 
-    if(done) return;
-    done = TRUE;
-    iocshRegister(&caPutLogInitDef,caPutLogInitCall);
-    iocshRegister(&caPutLogReconfDef,caPutLogReconfCall);
-    iocshRegister(&caPutLogShowDef,caPutLogShowCall);
-    iocshRegister(&caPutLogSetTimeFmtDef,caPutLogSetTimeFmtCall);
+    switch (caPutLogRegisterDone) {
+    case 0:
+        iocshRegister(&caPutLogInitDef,caPutLogInitCall);
+        iocshRegister(&caPutLogReconfDef,caPutLogReconfCall);
+        iocshRegister(&caPutLogShowDef,caPutLogShowCall);
+        iocshRegister(&caPutLogSetTimeFmtDef,caPutLogSetTimeFmtCall);
+        iocshRegister(&caPutJsonLogInitDef,caPutJsonLogInitCall);
+        caPutLogRegisterDone = 1;
+        break;
+
+    case 1:
+        /* Second registration, no problem */
+        break;
+
+    case 2:
+        errlogPrintf(ERL_WARNING
+            ": Registration of caPutLog commands skipped, as\n"
+            "  the caPutJsonLog commands have already been registered.\n"
+            "  This IOC may have been built with both 'caPutJsonLog.dbd'\n"
+            "  and 'caPutLog.dbd', please rebuild it using only one.\n");
+        break;
+    }
 }
 epicsExportRegistrar(caPutLogRegister);